requirements.txt
setuptools
numpy
pandas
matplotlib
lightgbm
tqdm
plotly
pyarrow
polars
seaborn
shap
ipywidgets
xgboost
openpyxl
holidays
pytorch